+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/feature/legalhold/MembersHavingLegalHoldClientUseCaseTest.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.kalium.logic.feature.conversation
+package com.wire.kalium.logic.feature.legalhold
 
 import com.wire.kalium.logic.StorageFailure
 import com.wire.kalium.logic.data.client.Client
